Introduced yaml, road delete endpoint, refactored a lot, readme updated, json validate almost done, db updated

+++ src/main/java/com/snk/starkkrumm/model/RoadRequest.java
@@ -15,7 +15,7 @@ public class RoadRequest {
     private String date;
     private Integer distanceBig;
     private Integer distanceSmall;
-    private Integer consumption1;
-    private Integer consumption2;
-    private Integer consumption3;
+    private Double consumption1;
+    private Double consumption2;
+    private Double consumption3;
 }
